# user-service/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Swagger for backend API service\nGet the Bearer token on the Authentication Service\nJSON Link: \u003ca href=/swagger/doc.json\u003edocs.json\u003c/a\u003e",
        "title": "Swagger for go-service",
        "contact": {},
        "license": {},
        "version": "1.0"
    },
    "paths": {
...
        "/order/check-out": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Endpoint to Insert an Order with status checkout will need shop_id and list of product_id and their quantity",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "order"
                ],
                "summary": "checkout an Order",
                "parameters": [
                    {
                        "description": "Add Product",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.CheckoutOrderRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.CheckoutOrderResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.CheckoutOrderRequest": {
            "type": "object",
            "properties": {
                "products": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.ProductCheckoutOrderRequest"
                    }
                },
                "shop_id": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "model.CheckoutOrderResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "order_id": {
                    "type": "integer"
                }
            }
        },
        "model.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "model.ProductCheckoutOrderRequest": {
            "type": "object",
            "properties": {
                "product_id": {
                    "type": "integer",
                    "example": 1
                },
                "quanttity": {
                    "type": "integer",
                    "example": 1
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}